Do not check for `config set` variable changes in config.bash:zz_variable
Doing so makes it impossible to call `prozzie config set` on setup to
change variables that always returns the same, like the monitor mibs
one. `zz_variable` calls config set (so, it calls _sanitize) on these
variables twice, but this is OK at this moment, all the sanitize
functions are idempotent.