Fix pexpect coverage in test
BASH_XTRACEFD actual file descriptor was not inherited on pexpect
spawned tty. That means that coverage information was not produced to
kcov, so it weas basically lost. At the end, it can cause even a false
OK, since bash exports coverage information via stderr in case that
cannot open fd, so pexpect was obtaining that debug output.

The patch to pexpect is not available in any release, so it's currently
bound to the prozzie repo until mainline pexpect apply it.